+++ openfisca_mali/variables/caracteristiques_socio_demographiques.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-
-
 from openfisca_core.model_api import *
 from openfisca_mali.entities import *